make sure solid files are read fully.
The v29 decoder was not necessarily reading the end of file marker.
Because the limit reader would stop once all data had been read it
may not request the end of file marker (which may also tell decoder
to read new decoder information for next file).
 Now save the decoder reader for solid files and read it fully before
starting on the next file.