make fileBlockReader's next() return errArchiveEnd instead of io.EOF at end block
Need to know whether the archive volume ended due to reading an end
block, or we really reached the end of the file (without finding an end
block).
This is for a later change that will allow multi volume archives with
volume files that don't have end blocks. It also makes the error more
explicit if an archive ends mid packed file.
packedFileReader() converts the errArchiveEnd to an io.EOF to preserve
the current behaviour of Next() returning io.EOF at archive end.