remove reset field from filterBlock
It was only used by V3 filters, and it was really just used for clearing
the cache of vm filters and their code. In theory maybe a newer filter
could cancel a previous one, causing it not to be run. But im not sure
its possible in the C++ code, or that it would even occur in the real
world. It also adds complexity in processing filters, in that filling
the decode window to get enough data for a filter could also cancel it.
Removing it so I can simplify the filter decoding.