allow for case for ppm where rarBitReader may have cached bits.
If a PPM block starts and there are more than 8 cached bits in
rarBitReader then some bytes may be ignored. This would only occur if
there is a single file which combines PPM and LZ blocks, and a previous
LZ block has a large enough number of cached bits.
To read the cached bytes, unpack the bytes and then create a new
replaceBitReader. When bytes() is called after the unpacked bytes have
been consumed, rarBitReader will now be in same state as before the bits
were unpacked. This ensures the cached bytes aren't skipped.
This will allow rarBitReader to more aggressively cache bits to improve
the lz decoder performance.