fix possible infinite loop with v5 on 32bit platforms
Bit reader for rar v5 may get into an infinite loop if called with a
value > integer size - 8 bits. There is a small chance for this to
happen on 32bit platforms in one case where readBits() is called. Fix
this case for 32bit platforms by splitting the read if it is too large.
More general fixes attempted on the bit reader had negative performance
impacts (eg. changing from int to int64).