change cipherBlockReader to use a byteReader
Changed cipherBlockReader from taking a packedFileReader to taking a
byteReader to read input from.
cipherBlockReader was the only reader that used the blocks() method, so
we can move the block calculation logic into where it is used. And it
will make cipherBlockReader more consistant with the other readers that
take a byteReader.
A future change will remove the blocks() method from packedFileReader
and blockReader.